backend/requirements.txt
django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9  # For PostgreSQL support
